refactor(ui,config): single-source duplicated helpers and flatten theme package
- LocationConfigScreen and LocationAppComponents each carried a private
  copy of parseDouble/parseFloat/parseInt plus a differently-named copy of
  the same result type. The copies had diverged: the screen's parseFloat
  tested `value < min || value > max`, which is false for NaN, and "NaN"
  parses fine via toFloatOrNull -- so it accepted NaN where the per-app
  editor rejected it. (ConfigRegistry.validateLocation catches it later via
  isFinite, so this was a UI-layer inconsistency rather than a stored-value
  bug.) Extracted to LocationValueParsing.kt using the NaN-rejecting range
  check throughout.

- SUPPORTED_LANGUAGES was declared identically in ConfigRegistry and
  AppPreferences; ConfigRegistry now reads the one in AppPreferences.
  Left "ko" in place rather than removing it: it is offered by the
  language_codes arrays in all three locales, but localeFilters ships only
  en/zh-rCN/ja and there is no values-ko, so choosing Korean silently falls
  back to default strings. Whether to translate or drop it is a product
  call, so it is documented at the constant instead of decided here.

- ui/ui/theme (an unflattened Studio template package) -> ui/theme, across
  16 files whose imports read asia.nana7mi.arirang.ui.ui.theme.

- ClipboardDialogActivity.kt renamed to ConfirmDialogActivity.kt, the class
  it actually declares and the name the manifest registers.

Verified: debug + release + selfcheck build, 72 tests green.
