+++ app/src/main/java/asia/nana7mi/arirang/ui/activity/SensorListActivity.kt
@@ -8,7 +8,7 @@ import androidx.activity.enableEdgeToEdge
 import asia.nana7mi.arirang.R
 import asia.nana7mi.arirang.data.datastore.SensorConfigPrefs
 import asia.nana7mi.arirang.ui.screen.sensor.SensorListScreen
-import asia.nana7mi.arirang.ui.ui.theme.ArirangTheme
+import asia.nana7mi.arirang.ui.theme.ArirangTheme
 
 class SensorListActivity : ComponentActivity() {
