+++ app/src/main/java/asia/nana7mi/arirang/ui/fragment/AboutFragment.kt
@@ -8,7 +8,7 @@ import androidx.compose.ui.platform.ComposeView
 import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import asia.nana7mi.arirang.ui.screen.about.AboutScreen
-import asia.nana7mi.arirang.ui.ui.theme.ArirangTheme
+import asia.nana7mi.arirang.ui.theme.ArirangTheme
 
 class AboutFragment : Fragment() {
