+++ app/src/main/java/asia/nana7mi/arirang/ui/fragment/SettingsFragment.kt
@@ -10,7 +10,7 @@ import androidx.compose.ui.platform.ViewCompositionStrategy
 import androidx.fragment.app.Fragment
 import asia.nana7mi.arirang.ui.activity.AdvancedSettingsActivity
 import asia.nana7mi.arirang.ui.screen.settings.SettingsScreen
-import asia.nana7mi.arirang.ui.ui.theme.ArirangTheme
+import asia.nana7mi.arirang.ui.theme.ArirangTheme
 
 class SettingsFragment : Fragment() {
